Add in built target to force serial build
A parallel build does not work. If a user installs using octave's `pkg
build` command, it seems to attempt a parallel build using the Makefile,
and this fails. So, this commit adds a directive to ensure that this
makefile is processes serially, even if the -jX options are passed
to make.